requirements.txt
django==3.2
django-filter==2.4.0
django-environ==0.4.5
drf-yasg==1.20.0
djangorestframework==3.12.4
postgres==3.0.0
gunicorn==20.1.0
